cleanup & continue on Stargate serve support
